scraper/requirements.txt
selenium>=4.1.0
PyYAML>=6.0
aiohttp>=3.8.1
aiofiles>=0.8.0
redis>=4.3.1
